Internalize and functionalize observer
