+++ logger_test.go
@@ -25,6 +25,7 @@ import (
 	"testing"
 
 	"go.uber.org/zap/internal/exit"
+	"go.uber.org/zap/internal/observer"
 	"go.uber.org/zap/testutils"
 	"go.uber.org/zap/zapcore"
 
@@ -35,7 +36,7 @@ func TestLoggerDynamicLevel(t *testing.T) {
 	// Test that the DynamicLevel applys to all ancestors and descendants.
 	dl := DynamicLevel()
 
-	withLogger(t, dl, nil, func(grandparent Logger, _ *zapcore.ObservedLogs) {
+	withLogger(t, dl, nil, func(grandparent Logger, _ *observer.ObservedLogs) {
 		parent := grandparent.With(Int("generation", 1))
 		child := parent.With(Int("generation", 2))
 
@@ -72,11 +73,11 @@ func TestLoggerDynamicLevel(t *testing.T) {
 
 func TestLoggerInitialFields(t *testing.T) {
 	fieldOpts := opts(Fields(Int("foo", 42), String("bar", "baz")))
-	withLogger(t, DebugLevel, fieldOpts, func(logger Logger, logs *zapcore.ObservedLogs) {
+	withLogger(t, DebugLevel, fieldOpts, func(logger Logger, logs *observer.ObservedLogs) {
 		logger.Info("")
 		assert.Equal(
 			t,
-			zapcore.ObservedLog{Context: []zapcore.Field{Int("foo", 42), String("bar", "baz")}},
+			observer.LoggedEntry{Context: []zapcore.Field{Int("foo", 42), String("bar", "baz")}},
 			logs.AllUntimed()[0],
 			"Unexpected output with initial fields set.",
 		)
@@ -85,14 +86,14 @@ func TestLoggerInitialFields(t *testing.T) {
 
 func TestLoggerWith(t *testing.T) {
 	fieldOpts := opts(Fields(Int("foo", 42)))
-	withLogger(t, DebugLevel, fieldOpts, func(logger Logger, logs *zapcore.ObservedLogs) {
+	withLogger(t, DebugLevel, fieldOpts, func(logger Logger, logs *observer.ObservedLogs) {
 		// Child loggers should have copy-on-write semantics, so two children
 		// shouldn't stomp on each other's fields or affect the parent's fields.
 		logger.With(String("one", "two")).Info("")
 		logger.With(String("three", "four")).Info("")
 		logger.Info("")
 
-		assert.Equal(t, []zapcore.ObservedLog{
+		assert.Equal(t, []observer.LoggedEntry{
 			{Context: []zapcore.Field{Int("foo", 42), String("one", "two")}},
 			{Context: []zapcore.Field{Int("foo", 42), String("three", "four")}},
 			{Context: []zapcore.Field{Int("foo", 42)}},
@@ -109,7 +110,7 @@ func TestLoggerLogPanic(t *testing.T) {
 		{func(logger Logger) { logger.Check(PanicLevel, "bar").Write() }, true, "bar"},
 		{func(logger Logger) { logger.Panic("baz") }, true, "baz"},
 	} {
-		withLogger(t, DebugLevel, nil, func(logger Logger, logs *zapcore.ObservedLogs) {
+		withLogger(t, DebugLevel, nil, func(logger Logger, logs *observer.ObservedLogs) {
 			if tt.should {
 				assert.Panics(t, func() { tt.do(logger) }, "Expected panic")
 			} else {
@@ -137,7 +138,7 @@ func TestLoggerLogFatal(t *testing.T) {
 		{func(logger Logger) { logger.Check(FatalLevel, "bar").Write() }, "bar"},
 		{func(logger Logger) { logger.Fatal("baz") }, "baz"},
 	} {
-		withLogger(t, DebugLevel, nil, func(logger Logger, logs *zapcore.ObservedLogs) {
+		withLogger(t, DebugLevel, nil, func(logger Logger, logs *observer.ObservedLogs) {
 			stub := exit.WithStub(func() {
 				tt.do(logger)
 			})
@@ -156,7 +157,7 @@ func TestLoggerLogFatal(t *testing.T) {
 }
 
 func TestLoggerLeveledMethods(t *testing.T) {
-	withLogger(t, DebugLevel, nil, func(logger Logger, logs *zapcore.ObservedLogs) {
+	withLogger(t, DebugLevel, nil, func(logger Logger, logs *observer.ObservedLogs) {
 		tests := []struct {
 			method        func(string, ...zapcore.Field)
 			expectedLevel zapcore.Level
@@ -184,7 +185,7 @@ func TestLoggerLeveledMethods(t *testing.T) {
 func TestLoggerAlwaysPanics(t *testing.T) {
 	// Users can disable writing out panic-level logs, but calls to logger.Panic()
 	// should still call panic().
-	withLogger(t, FatalLevel, nil, func(logger Logger, logs *zapcore.ObservedLogs) {
+	withLogger(t, FatalLevel, nil, func(logger Logger, logs *observer.ObservedLogs) {
 		msg := "Even if output is disabled, logger.Panic should always panic."
 		assert.Panics(t, func() { logger.Panic("foo") }, msg)
 		assert.Panics(t, func() {
@@ -199,7 +200,7 @@ func TestLoggerAlwaysPanics(t *testing.T) {
 func TestLoggerAlwaysFatals(t *testing.T) {
 	// Users can disable writing out fatal-level logs, but calls to logger.Fatal()
 	// should still terminate the process.
-	withLogger(t, FatalLevel+1, nil, func(logger Logger, logs *zapcore.ObservedLogs) {
+	withLogger(t, FatalLevel+1, nil, func(logger Logger, logs *observer.ObservedLogs) {
 		stub := exit.WithStub(func() { logger.Fatal("") })
 		assert.True(t, stub.Exited, "Expected calls to logger.Fatal to terminate process.")
 
@@ -215,20 +216,20 @@ func TestLoggerAlwaysFatals(t *testing.T) {
 }
 
 func TestLoggerDPanic(t *testing.T) {
-	withLogger(t, DebugLevel, nil, func(logger Logger, logs *zapcore.ObservedLogs) {
+	withLogger(t, DebugLevel, nil, func(logger Logger, logs *observer.ObservedLogs) {
 		assert.NotPanics(t, func() { logger.DPanic("") })
 		assert.Equal(
 			t,
-			[]zapcore.ObservedLog{{Entry: zapcore.Entry{Level: DPanicLevel}, Context: []zapcore.Field{}}},
+			[]observer.LoggedEntry{{Entry: zapcore.Entry{Level: DPanicLevel}, Context: []zapcore.Field{}}},
 			logs.AllUntimed(),
 			"Unexpected log output from DPanic in production mode.",
 		)
 	})
-	withLogger(t, DebugLevel, opts(Development()), func(logger Logger, logs *zapcore.ObservedLogs) {
+	withLogger(t, DebugLevel, opts(Development()), func(logger Logger, logs *observer.ObservedLogs) {
 		assert.Panics(t, func() { logger.DPanic("") })
 		assert.Equal(
 			t,
-			[]zapcore.ObservedLog{{Entry: zapcore.Entry{Level: DPanicLevel}, Context: []zapcore.Field{}}},
+			[]observer.LoggedEntry{{Entry: zapcore.Entry{Level: DPanicLevel}, Context: []zapcore.Field{}}},
 			logs.AllUntimed(),
 			"Unexpected log output from DPanic in development mode.",
 		)
@@ -236,11 +237,11 @@ func TestLoggerDPanic(t *testing.T) {
 }
 
 func TestLoggerNoOpsDisabledLevels(t *testing.T) {
-	withLogger(t, WarnLevel, nil, func(logger Logger, logs *zapcore.ObservedLogs) {
+	withLogger(t, WarnLevel, nil, func(logger Logger, logs *observer.ObservedLogs) {
 		logger.Info("silence!")
 		assert.Equal(
 			t,
-			[]zapcore.ObservedLog{},
+			[]observer.LoggedEntry{},
 			logs.AllUntimed(),
 			"Expected logging at a disabled level to produce no output.",
 		)
@@ -275,7 +276,7 @@ func TestLoggerAddCaller(t *testing.T) {
 		{opts(AddCaller(), AddCallerSkip(1), AddCallerSkip(3)), `.+/src/runtime/.*:[\d]+$`},
 	}
 	for _, tt := range tests {
-		withLogger(t, DebugLevel, tt.options, func(logger Logger, logs *zapcore.ObservedLogs) {
+		withLogger(t, DebugLevel, tt.options, func(logger Logger, logs *observer.ObservedLogs) {
 			logger.Info("")
 			output := logs.AllUntimed()
 			assert.Equal(t, 1, len(output), "Unexpected number of logs written out.")
@@ -291,7 +292,7 @@ func TestLoggerAddCaller(t *testing.T) {
 
 func TestLoggerAddCallerFail(t *testing.T) {
 	errBuf := &testutils.Buffer{}
-	withLogger(t, DebugLevel, opts(AddCaller(), ErrorOutput(errBuf)), func(log Logger, logs *zapcore.ObservedLogs) {
+	withLogger(t, DebugLevel, opts(AddCaller(), ErrorOutput(errBuf)), func(log Logger, logs *observer.ObservedLogs) {
 		// TODO: Use AddCallerSkip
 		logImpl := log.(*logger)
 		logImpl.callerSkip = 1e3
@@ -312,11 +313,11 @@ func TestLoggerAddCallerFail(t *testing.T) {
 }
 
 func TestLoggerAddStacks(t *testing.T) {
-	assertHasStack := func(t testing.TB, obs zapcore.ObservedLog) {
+	assertHasStack := func(t testing.TB, obs observer.LoggedEntry) {
 		assert.Contains(t, obs.Entry.Stack, "zap.TestLoggerAddStacks", "Expected to find test function in stacktrace.")
 	}
 
-	withLogger(t, DebugLevel, opts(AddStacks(InfoLevel)), func(logger Logger, logs *zapcore.ObservedLogs) {
+	withLogger(t, DebugLevel, opts(AddStacks(InfoLevel)), func(logger Logger, logs *observer.ObservedLogs) {
 		logger.Debug("")
 		assert.Empty(
 			t,
@@ -333,7 +334,7 @@ func TestLoggerAddStacks(t *testing.T) {
 }
 
 func TestLoggerConcurrent(t *testing.T) {
-	withLogger(t, DebugLevel, nil, func(logger Logger, logs *zapcore.ObservedLogs) {
+	withLogger(t, DebugLevel, nil, func(logger Logger, logs *observer.ObservedLogs) {
 		child := logger.With(String("foo", "bar"))
 
 		wg := &sync.WaitGroup{}
@@ -351,7 +352,7 @@ func TestLoggerConcurrent(t *testing.T) {
 		for _, obs := range logs.AllUntimed() {
 			assert.Equal(
 				t,
-				zapcore.ObservedLog{
+				observer.LoggedEntry{
 					Entry:   zapcore.Entry{Level: InfoLevel},
 					Context: []zapcore.Field{String("foo", "bar")},
 				},
