+++ zapcore/facility.go
@@ -95,112 +95,6 @@ func (iof *ioFacility) clone() *ioFacility {
 	}
 }
 
-// An ObservedLog is an encoding-agnostic representation of a log message.
-type ObservedLog struct {
-	Entry
-	Context []Field
-}
-
-func (o ObservedLog) untimed() ObservedLog {
-	o.Entry.Time = time.Time{}
-	return o
-}
-
-// ObservedLogs is a concurrency-safe, ordered collection of observed logs. It
-// has a fixed capacity, after which new logs begin overwriting the oldest
-// observations.
-type ObservedLogs struct {
-	cap  int
-	mu   sync.RWMutex
-	logs []ObservedLog
-}
-
-// Add appends a new observed log to the collection.
-func (o *ObservedLogs) Add(ent Entry, context []Field) {
-	log := ObservedLog{
-		Entry:   ent,
-		Context: context,
-	}
-	o.mu.Lock()
-	o.logs = append(o.logs, log)
-	if len(o.logs) > o.cap {
-		o.logs = o.logs[1:]
-	}
-	o.mu.Unlock()
-}
-
-// Len returns the number of items in the collection.
-func (o *ObservedLogs) Len() int {
-	o.mu.RLock()
-	defer o.mu.RUnlock()
-	return len(o.logs)
-}
-
-// All returns a copy of all the observed logs.
-func (o *ObservedLogs) All() []ObservedLog {
-	o.mu.RLock()
-	ret := make([]ObservedLog, len(o.logs))
-	for i := range o.logs {
-		ret[i] = o.logs[i]
-	}
-	o.mu.RUnlock()
-	return ret
-}
-
-// AllUntimed returns a copy of all the observed logs, but overwrites the
-// observed timestamps with time.Time's zero value. This is useful when making
-// assertions in tests.
-func (o *ObservedLogs) AllUntimed() []ObservedLog {
-	o.mu.RLock()
-	ret := make([]ObservedLog, len(o.logs))
-	for i := range o.logs {
-		ret[i] = o.logs[i].untimed()
-	}
-	o.mu.RUnlock()
-	return ret
-}
-
-type observerFacility struct {
-	LevelEnabler
-	sink    *ObservedLogs
-	context []Field
-}
-
-// NewObserver creates a new facility that buffers logs in memory (without any
-// encoding). It's particularly useful in tests, though it can serve a variety
-// of other purposes as well. This constructor returns the facility itself and
-// a function to retrieve the observed logs.
-func NewObserver(enab LevelEnabler, cap int) (Facility, *ObservedLogs) {
-	fac := &observerFacility{
-		LevelEnabler: enab,
-		sink:         &ObservedLogs{cap: cap, logs: make([]ObservedLog, 0, cap)},
-	}
-	return fac, fac.sink
-}
-
-func (o *observerFacility) With(fields []Field) Facility {
-	return &observerFacility{
-		LevelEnabler: o.LevelEnabler,
-		sink:         o.sink,
-		context:      append(o.context[:len(o.context):len(o.context)], fields...),
-	}
-}
-
-func (o *observerFacility) Write(ent Entry, fields []Field) error {
-	all := make([]Field, 0, len(fields)+len(o.context))
-	all = append(all, o.context...)
-	all = append(all, fields...)
-	o.sink.Add(ent, all)
-	return nil
-}
-
-func (o *observerFacility) Check(ent Entry, ce *CheckedEntry) *CheckedEntry {
-	if o.Enabled(ent.Level) {
-		ce = ce.AddFacility(ent, o)
-	}
-	return ce
-}
-
 type counters struct {
 	sync.RWMutex
 	counts map[string]*atomic.Uint64
